% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noter.R
\name{noter}
\alias{noter}
\title{noter}
\usage{
noter(
  code,
  comments = NULL,
  taskID = tmpID(),
  path = getOption("noter_path"),
  check = getOption("noter_check"),
  template = getOption("noter_template")
)
}
\description{
noter
}
